+++ Cargo.toml
@@ -6,7 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-anyhow = "1.0.75"
 axum = "0.6.20"
 config = "0.13.3"
 dotenvy = "0.15.7"
